requirements.txt
appnope==0.1.0
appscript==1.0.1
CacheControl==0.12.3
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
decorator==4.1.2
idna==2.5
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
msgpack-python==0.4.8
numpy==1.13.1
pandas==0.20.3
pandas-datareader==0.5.0
pandasql==0.7.3
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psutil==5.2.2
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
python-dotenv==0.6.4
pytz==2017.2
requests==2.18.3
requests-file==1.4.2
requests-ftp==0.3.1
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.12
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
xlwings==0.11.4
